Change command line behavior of --read-json and --save-json.
The --read-json flag instead of being a simple boolean flag, now takes a
directory path as its argument so that json files
saved to a subdirectory from a previous run can more easily be reused.

The --save-json option becomes --no-save-json, so the default is now to save
json files by default.